requirements.txt
beautifulsoup4==4.12.3
PyExecJS==1.5.1
Requests==2.32.3
